.java file changes:
* Set the class as final.
* Made Logger, LayoutXmlCache, LayoutBean, and ChannelCache all final.
* Set layout.jsp to threaded and increase its buffer to 20k.
* Fixed bug in LayoutBean when the getChannelInstance(String) method is
called. It used to just return a null, now it will try to find the channel in
the layout and return something.
* Fixed a problem in the DispatchBean where, if a channel gives its
channelID twice, it would concat the two Ids together with a comma.
* Minor fix to ubc_authentication.jsp to protect against user name being null.
* The setting of the portal base directory is now done in the
GenericPortalBean and the initialize() method is called by all JSPs, so
that any of the JSPs can be first accessed after the server is started.
* Really fixed the problem with ubc_authentication.jsp this time.
* Changed findLayoutInstance so that when the guest user LayoutBean is
first created, the getLayoutXml function of layoutBean is called to initialize
the guest layout bean properly.  This was needed so that
ubc_authentication.jsp would be properly initialized.
* Changes to name a Tab "Blank tab" if the user leaves the Tab name
blank (i.e. null)
* Fixed the tab naming problem.
* Added the UtilitiesBean.removeSpecialChars methods
* Fixed the "revert to default layout" functionality
* Made the RdbmServices object static so that it is not created multiple times.
* Until the portal base directory isn't set, logging will be to the console.
That way the Logger will start functioning normally after the first JSP is
accessed.
* Suppressing the stack trace in the logs when previewChannel() catches an
IOException due to Broken pipes.
* Changed cache control settings back to what they were in 1.0 to accomodate
problems with Internet Explorer 5.0
* Suppressing the stack trace in the logs when renderXML() catches an
IOException due to Broken pipes.
* CGenericXSLST.java added from UBC code.

.jsp file changes:

* The setting of the portal base directory is now done in the
GenericPortalBean and the initialize() method is called by all JSPs, so that
any of the JSPs can be first accessed after the server is started.
* Removed all content from index.jsp and replaced it with an include of layout.jsp.
* Fixed bug in ubc_authentication.jsp where it was messing up the layout if
you went to the ubc_authentication.jsp before going to index.jsp.  This was
because the LayoutBean does some initialization stuff when it is first
loaded, but ubc_authentication.jsp did not create the LayoutBean.
* Fixed a bug in the personalizeColor.jsp where the <select> options were
not encoding spaces to + for the names of the color types.
* Modification to ensure naming a Tab "Blank tab" if the user leaves the
Tab name blank (i.e. null) or enters non-alphanumeric characters
* Added a JavaScript alert if the user has entered character combinations
for tab names that ends up with a "Blank Tab".


git-svn-id: https://source.jasig.org/uPortal/branches/rel-1-0-patches@4339 f5dbab47-78f9-eb45-b975-e544023573eb