Added UserPreferences folder to the guest user's layout
Updated guest users channels to make them removable, they will still not be removable because their parent folder is immutable

Added guest user to the UP_PERSON_DIR table so that the guest user can log in, this is commented out by default to prevent someone from setting up a demo without realizing that anyone can log in as guest. Uncomment the row and CHANGE THE PASSWORD if you want to be able to log in as the guest user to change the layout.


git-svn-id: https://source.jasig.org/uPortal/trunk@5231 f5dbab47-78f9-eb45-b975-e544023573eb