AdditionalDescriptors that implement the Map interface are now handled.
AdditionalDescriptors that implement the IPerson interface are handled differently. The attributes are copied into the IPerson provided by the caller.

There is a new flag in portal.properties for using PersonDirectory or not.

setFullName() is called at the end of the method, by then the IPerson object should have been completely populated.


git-svn-id: https://source.jasig.org/uPortal/trunk@5248 f5dbab47-78f9-eb45-b975-e544023573eb