UP-869 UP-874
Updated AuthorizationImpl to use the new CacheFactory which uses WhirlyCache by default.  Also modified AuthorizationImpl to not doing a map copy on the creation of a new principal.

As a result of these two JIRA issues, a generic CacheFactory was created (addition of 3 classes, 3 jars, and some config files).  This cachefactory can be used by other parts of the Portal as needed.


git-svn-id: https://source.jasig.org/uPortal/trunk@10134 f5dbab47-78f9-eb45-b975-e544023573eb