UP-1798 Replacing changes that were accidentally removed with some branch work.  This work is documented in UP-1798.  The original change comment is below.
Removed the System.out.println version of the stats event handler. It was redundant with the log4j version.

Modified the statsContext.xml to list the major event types (everything that directly subclasses PortalEvent), commented out all events except the user login/logout and session created/destroyed events. Modifying the list of events that are desired is now, simply a matter of customizing the lines that are commented/uncommented in statsContext.xml.


git-svn-id: https://source.jasig.org/uPortal/trunk@15404 f5dbab47-78f9-eb45-b975-e544023573eb