UP-3251:  issue importing layouts in the 4.0.2 build with data from 3.2.x build.
The problem is that the import process was converting bad DLM pathrefs
(representing orphaned data) to zero-length strings.  It seems Oracle, for no
reason I can come to terms with, interprets zero-length strings as null on
insert/update.  The column in question is non-nullable, and this combination
gives rise to the error.