Fixed improper anchor nesting
Fluid's inline edit component was generating an anchor tag inside the anchor tag for the tab, which is a big w3c no-no. Since the component did not have a option for the markup, I copied the source function into up-tab-mange.js and made the change directly. It now generates a span tag for the button instead of an anchor.