UP-3477 Switch JpaStatisticalSummary to an embedded object
To reduce the complexity of the table structure convert
JpaStatisticalSummary into an embedded object so its columns will simply
be part of the parent object.

Add orm.xml to map the commons-math stats objects used in
JpaStatisticalSummary. This bypasses the need for serialization to JSON
and reduces the IO overhead of tracking statistics.