NOJIRA. Update licensing NOTICE. And pointed at SSP-O-S-P v1.1.0-SNAPSHOT.
Conflicts:
	NOTICE

Resolved NOTICE conflict by regenerating that file. Upgraded the SSP-O-S-P dependencies
because now that SPC is deploying 1.1.0-SNAPSHOT, 1.0.0-SNAPSHOT is effectively orphaned
and there should be no reason to continue to maintain that branch. This was also
convenient for this particular commit b/c it meant we didn't have to go back to
SSP-O-S-P 1.0.0-SNAPSHOT and add licensing metadata and/or configure additional
licensing metadata here to satisfy the NOTICE generator.