NOJIRA De-aggregate uportal-java-api and depend on release version.
Don't want to aggregate the uportal-java-api (platform-java-api) in
platform parent because the former should be released separately from the
latter, but if it's aggregated it prevents auto/bulk release of child
projects in the latter.

Depending on the latest release also simplifies most builds since you
don't need to build uportal-java-api first. Though this change mostly
affects projects external to platform-parent.