SSP-531 Add platform version and SSP version to footer
This processes pull request #149 and works around an incompatibility
with the deploy-ear target, which, because of a bug in in
uportal-ant-tasks which skips MANIFEST.MF when deploying war files.
The workaround is to re-deploy the war with "fastWarDeploy" after
deploy-ear does it's normal work.

I also removed an unused uportal-war/pom.xml property from the
original patch in the pull request and fixed up some whitespace.
Thanks, Paul!