UP-3811 Add merging person dir DAO and clarify docs
Don't want to add custom attribute sources to the list of DAOs in
the cascading DAO. I.e. the existing comments in
personDirectoryContext.xml gave one the wrong idea.

This commit wraps the cascading DAO in a merging DAO and updates
comments to strongly suggest adding custom sources to the latter
rather than the former.

This is a cherry-pick and re-word of
67612e9794051fc3cfa7d0303f6b85b8a140fd9c from ssp-master.