SSP-1669 Configurable persons search wildcarding exclusions
This brings parity with recent work on SSP's PersonDirectory fork
(PERSONDIR-77). Can now set a property in portal.properties to list
the local uP attributes which should not be automatically wrapped in
wildcards when searched on.

This is important for SSP because whenever we search on the SSP_ROLES
attribute we almost always only want exact matches.

Ideally this behavior would be moved "up" into the queries themselves,
because we don't necessarily want wildcarding shut off in all cases.
But this sort of DAO-layer wildcarding and predicate logic enforcement
is common in PD as well (e.g. you specify search term ANDing or ORing
at the DAO layer, not the query layer). And if you really need to, you
can always include explicit wildcards in the search term and those
*will* be honored, regardless of the wildcard exclusion config. (In
SSP you typically wouldn't include explicit wildcards, though, because
you're often searching against multi-valued LDAP attributes, for which
wildcards won't work.)