SSP-1888 Remove assertion that requires SSP_CONFIGDIR to be unset
Looks like we don't actually need this because the default property
source ordering in StandardEnvironment is different than what I
thought... JVM properties are preferred over system env vars. So
as long as tests always set the location of SSP_CONFIGDIR as a
system property, our tests should be fine... which they do seem to
be.