UP-3926 - Respondr skin refactor initial restructure of .less files
refactoring of the skins.  moved things from uPortal4 to defaultSkin and also moved the majority of the styles to Common.  Including the bootstrap.less files as well instead of including them from the ResourceServer

Fixed the build time issue with the Less files not being compiled.  Added a more up to date version of less.js.  Also fixed an incorrect path to the jQuery UI Smoothness theme file.  Added *.sublime* to the .gitignore file.  Changed the path for the background image

set the color of the tab edit field to @grayscale10.  Corrected the paths on the images in the gallary.less so they no longer give 404s