UP-4010 - Respondr View changes (mobile, emergency portlet, etc)
When in mobile view, some of the elements don't align up properly. Restructured some of the code to be "mobile first".

Added a background to the emergency alert notification portlet. This allows various skins to set it accordingly.

Hide the tips and customize draw button when viewing on a small device.

Added variable for setting the logout link color.