Remove the Calendar portlet.
Drops dependency on calendar portlet 2.1.3-M4, removing it from the pom.xml, dropping the overlay upon it, and removing the portlet publication and layout reference to it.  This allows the uPortal build to succeed despite the calendar portlet's caldav4j dependency not yet being available in Central and no longer being available in third party repo.