UP-4192 CVE-2014-5059 Restore SSP CAS disablement
Previous UP-4192 patch to security.properties was integrated into
SSP source mostly as-is with just a few fix-ups for Git conflicts
so we have a reasonable baseline against which to track
truly SSP-specific customizations, of which this commit is one.

In particular it re-disabled the CAS security context. It also
fixes a minor typo in a comment.