NOJIRA: purpose of a French translation
It's a first purpose, and I can change some text.
It comes from some ESUP translations, but not all, somes like "Principals" wasn't adapted and I prefer to doesn't translate "Principals" as it's diffciult to find something related to permissions in french.
More in Unix world we are familiar with Principals.

In French rules are :
- we should add a No-Break Space before each : and ;
- a space before and after ! ? §