UP-3780: Providing some test cases
Provinding some test case, this will only check if json.xsl and json-v4.3.xsl are well formed.
I would loved to provide a test case on json format and with the bug, but there wasn't a simple test case on json done like it is with other xsl.
Or i didn't find it.