UP-4726: Configuration for proxyCAS url and for load-balancing portals
Providing a configuration for load-balanced portal using proxy-CAS feature. You need to have a one public servername access by portail with your common load-balanced domain to be able to target a specific portal.
In this case you need to set only the real hostname and context in the configuration to work.
This configuration is used by a lot of ESUP consortium members.