chore: limit concurrent renovate PRs
with the incoming fixes and improvements coming to renovate, it can
discover all of uPortal's dependencies and recommend updates.
On a test run on this, it opened 54 pull requests.
To avoid bombarding uPortal with far too many PRs, setting a limit on
how many PRs renovate will open.

This setting tells renovate to open up to 5 pull requests, and wait
until those requests are merged or declined before opening more to
replace the closed PR.